+++ project/ovp/frontend/ovp/utils/constant.ts
@@ -7,8 +7,8 @@
     4. 숫자 (0-9) 하나 이상 포함
     5. 특수문자 (예: ~, `, !, @, #, $, %, ^, (, ), _, -, =, {, }, [, ], |, ;, :, <, >, ,, ., ?, /) 하나 이상 포함
 * */
-export const pwRegex =
+export const PWREGEX =
   /^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[~`!@#$%^()_\-=\[\]{}|;:<>,.\/?]).{8,56}$/;
 
-export const pwRegexErrorMsg =
+export const PWREGEX_ERROR_MSG =
   "비밀번호는 최소 8자, 최대56자여야 하며 대문자(AZ), 소문자(az), 숫자, 특수문자(예:!,%,@ 또는 #등)를 하나 이상 포함해야 합니다.";
